requirements.txt
beautifulsoup4==4.3.2
cffi==1.1.2
characteristic==14.3.0
cryptography==0.9.1
cssselect==0.9.1
Django==1.8.2
django-appconf==1.0.1
djangorestframework==3.1.3
enum34==1.0.4
gnureadline==6.3.3
idna==2.0
ipaddress==1.0.7
ipython==3.1.0
lxml==3.4.4
praw==3.0.0
psycopg2==2.6.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.14
pyOpenSSL==0.15.1
queuelib==1.2.2
requests==2.7.0
Scrapy==1.0.0rc3
service-identity==14.0.0
six==1.9.0
sqlparse==0.1.15
Twisted==15.2.1
Unidecode==0.4.18
update-checker==0.11
w3lib==1.11.0
zope.interface==4.1.2
